pronas-suite-system/services/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiohttp==3.9.1
asyncio==3.4.3

# NLP e ML
transformers==4.36.0
torch==2.1.0
sentence-transformers==2.2.2
spacy==3.7.2
scikit-learn==1.3.2
xgboost==2.0.2
numpy==1.24.3
pandas==2.1.4

# OCR
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
Pillow==10.1.0

# Documentos
python-docx==1.1.0
openpyxl==3.1.2
pdfkit==1.0.0
reportlab==4.0.7
jinja2==3.1.2

# Banco de dados
sqlalchemy==2.0.23
asyncpg==0.29.0
redis==5.0.1
elasticsearch==8.11.0

# Monitoramento
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0

# Utilitários
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2